fix packaging of audio processor worklet
