+++ realtime-core/RealtimeWebSocket/index.ts
@@ -1,2 +1,3 @@
 export * from "./RealtimeWebSocketConnection";
 export * from "./RealtimeWebSocketMediaManager";
+export * from "./RealtimeWebsocketAudioProcessor.worklet";
